Listed the alignment_filter metadata attribute.
This attribute existed in iRODS for a long time. At the time
of writing it is used only for Illumina data. However, it is
rather generic, there is no reason to restrict it to Illumina
data.

The following values are currently in use: 'phix', 'human',
'yhuman', 'xahuman'. For Illumina data the value of the
this attribute is derived from the value of the 'subset'
attribute of the component. Most likely this attribute
pre-dates the era of compositions and components.